Remove redundant PersonServiceIT
